requirements.txt
# FastAPI e dependências
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Firebase/Firestore
google-cloud-firestore==2.13.1
firebase-admin==6.1.0

# OpenAI
openai==1.3.7

# Machine Learning e NLP
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
unidecode==1.3.7
rapidfuzz==3.5.2

# Utilitários
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Desenvolvimento e testes
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# CORS
fastapi-cors==0.0.6

# Logging
loguru==0.7.2

# Cache
cachetools==5.3.2